Add ready event and handling for cooldowns

+++ src/main/java/us/eunoians/mcrpg/ability/CooldownableAbility.java
@@ -1,5 +1,7 @@
 package us.eunoians.mcrpg.ability;
 
+import us.eunoians.mcrpg.api.event.ability.CooldownableAbilityActivateEvent;
+
 /**
  * This interface represents an {@link Ability} that should be put on cooldown after activation.
  * <p>
@@ -7,11 +9,11 @@ package us.eunoians.mcrpg.ability;
  * <p>
  * 1) The {@link Ability} implementation extends this ({@link CooldownableAbility}) interface
  * <p>
- * 2) There is an {@link us.eunoians.mcrpg.api.event.CooldownableAbilityActivateEvent} called for this {@link CooldownableAbility}s activation
+ * 2) There is an {@link CooldownableAbilityActivateEvent} called for this {@link CooldownableAbility}s activation
  * that adheres to the Bukkit/Spigot event calling order.
  * <p>
  * If both of these conditions are true, then McRPG will automatically handle putting abilities on cooldown (with lunar support!!!)
- * and cancelling {@link us.eunoians.mcrpg.api.event.CooldownableAbilityActivateEvent} calls if the {@link CooldownableAbility} is on cooldown.
+ * and cancelling {@link CooldownableAbilityActivateEvent} calls if the {@link CooldownableAbility} is on cooldown.
  * <p>
  * This means that the implementor doesn't need to care about any cooldown related information besides providing implementation to the
  * methods provided in this interface.
@@ -24,5 +26,5 @@ public interface CooldownableAbility extends Ability {
      * Gets the amount of time in seconds that this {@link CooldownableAbility} should be on cooldown for after activation
      * @return The postivie zero exclusive amount of time in seconds this {@link CooldownableAbility} should be on cooldown for after activation.
      */
-    public long getCooldownDuration();
+    public int getCooldownDuration();
 }
